requirements.txt
PyQt5>=5.15.0
numpy>=1.19.0
sounddevice>=0.4.1
soundfile>=0.10.3
requests>=2.25.1
tqdm>=4.61.0
keyboard>=0.13.5
pynput>=1.7.3
pyperclip>=1.8.2
packaging>=21.0
google-generativeai>=0.3.0 